src/test/kotlin/com/wuyangnju/mytxkv/cli/impl/cmdhandler/CountCmdHandlerTest.kt
package com.wuyangnju.mytxkv.cli.impl.cmdhandler

import com.wuyangnju.mytxkv.kv.TransactionalKeyValueStoreType
import com.wuyangnju.mytxkv.kv.createTransactionalKeyValueStore
import org.junit.jupiter.api.Assertions.assertFalse
import kotlin.test.BeforeTest
import kotlin.test.Test
import kotlin.test.assertEquals
import kotlin.test.assertNull
import kotlin.test.assertTrue

class CountCmdHandlerTest {
    private lateinit var cmdHandler: ICmdHandler

    @BeforeTest fun beforeTest() {
        cmdHandler = CountCmdHandler(createTransactionalKeyValueStore(TransactionalKeyValueStoreType.InMemory))
    }

    @Test fun testMatches() {
        assertTrue(cmdHandler.matches(listOf("count", "123")))
        assertFalse(cmdHandler.matches(listOf("foo")))
    }

    @Test fun testValidate() {
        assertTrue(cmdHandler.validate(listOf("count", "123")))
        assertFalse(cmdHandler.validate(listOf("count")))
    }

    @Test fun testAlertMsg() {
        assertNull(cmdHandler.alertMsg(listOf("count", "123")))
    }

    @Test fun testProcess() {
        assertEquals("0", cmdHandler.process(listOf("count", "123")))
    }
}